Remove the use of ircsprintf.
Stub out the use of the ircsprintf.c file and use the native
sprintf/snprintf functions. Also tell GCC that various functions are
actually printf-like functions in disguise, and fix up the resulting
warnings.

This should help the code work better on 64-bit platforms, even if it
makes us a few microseconds slower.

+++ src/m_services.c
@@ -230,8 +230,8 @@ int m_svsnick(aClient *cptr, aClient *sptr, int parc, char *parv[])
 #endif
     sendto_common_channels(acptr, ":%s NICK :%s", parv[1], newnick);
     add_history(acptr, 1);
-    sendto_serv_butone(NULL, ":%s NICK %s :%d", parv[1], newnick,
-		       acptr->tsinfo);
+    sendto_serv_butone(NULL, ":%s NICK %s :%ld", parv[1], newnick,
+		       (long)acptr->tsinfo);
     if(acptr->name[0]) 
     {
         del_from_client_hash_table(acptr->name, acptr);
